refactor(Hackathon): File Structure changed and removes duplicate files
