+++ src/modules/Dashboard/modules/Hackathon/pages/styles.css
@@ -187,3 +187,227 @@
 		font-size: .8rem;
 	}
 }
+
+/*
+! TODO: Need to move to main module.css
+*/
+
+.toolTipStyle {
+	background-color: var(--blue);
+	color: var(--White);
+	border-radius: var(--borderRadius);
+}
+
+.box {
+	background-color: transparent;
+	display: flex;
+	width: 100%;
+	gap: .5rem;
+	flex-wrap: wrap;
+}
+
+.box .card-component {
+	align-items: flex-start;
+	background-color: #ffffff;
+	border: 0.97px solid;
+	border-color: #dbdbdb;
+	border-radius: 16.13px;
+	display: flex;
+	flex-direction: column;
+	padding: 25.81px;
+}
+
+.box .card-component:hover {
+	box-shadow: 4px 4px 21px #0000001f;
+}
+
+.box .frame {
+	align-items: flex-start;
+	display: flex;
+	flex-direction: column;
+	gap: 19.36px;
+	position: relative;
+	width: fit-content;
+}
+
+.box .div {
+	align-items: flex-start;
+	display: flex;
+	gap: 85.5px;
+	position: relative;
+	width: fit-content;
+}
+
+.box .title {
+	height: 35.81px;
+	min-width: 147.9px;
+	position: relative;
+}
+
+.box .text-wrapper {
+	color: var(--blue);
+	font-family: "Plus Jakarta Sans-Bold", Helvetica;
+	font-size: 22.6px;
+	font-weight: 700;
+	text-align: left;
+	letter-spacing: 0;
+	line-height: normal;
+	white-space: nowrap;
+	width: 144px;
+}
+
+.box .text-wrapper-2 {
+	color: var(--black);
+	font-family: "Plus Jakarta Sans-Light", Helvetica;
+	font-size: 12.9px;
+	font-weight: 300;
+	left: 0;
+	letter-spacing: 0;
+	line-height: normal;
+	position: absolute;
+	top: 26px;
+	white-space: nowrap;
+}
+
+.box .shared {
+	height: 24.52px;
+	min-width: 61.95px;
+	position: relative;
+}
+
+.box .frame-2 {
+	align-items: flex-start;
+	display: flex;
+	gap: 12px;
+	position: relative;
+	width: 62px;
+}
+
+.box .group {
+	background-color: #e5ebff;
+	border-radius: 12.26px;
+	height: 24.52px;
+	margin-bottom: 10px;
+	min-width: 24.52px;
+	position: relative;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	cursor: pointer;
+}
+
+.box .group-2 {
+	height: 28.39px;
+	min-width: 102.66px;
+	position: relative;
+
+}
+
+.box .text-wrapper-3 {
+	color: #000000cc;
+	font-family: "Plus Jakarta Sans-Light", Helvetica;
+	font-size: 9.7px;
+	font-weight: 300;
+	letter-spacing: 0;
+	line-height: normal;
+	position: absolute;
+}
+
+.box .overlap-group {
+	border-radius: 5.16px;
+	height: 17px;
+	width: 48px;
+	position: relative;
+	top: 15px;
+	gap: 0.11rem;
+	display: flex;
+	align-items: center;
+	justify-content: start;
+}
+
+.box .text-wrapper-4 {
+	color: var(--grey);
+	font-family: "Plus Jakarta Sans-Regular", Helvetica;
+	font-size: 8px;
+	font-weight: 400;
+	letter-spacing: 0;
+	line-height: normal;
+	padding: 0.3rem .5rem;
+	border: 1px solid var(--grey);
+	border-radius: 0.32263rem;
+	white-space: nowrap;
+}
+
+.box .frame-3 {
+	align-items: center;
+	display: flex;
+	justify-content: space-between;
+	gap: 12.91px;
+	position: relative;
+	width: 295.54px;
+}
+
+.box .frame-4 {
+	align-items: flex-start;
+	display: flex;
+	gap: 12.09px;
+	height: 25.69px;
+	position: relative;
+	width: 150.35px;
+}
+
+.box .text-wrapper-small {
+	color: var(--black);
+	font-family: "Plus Jakarta Sans-Medium", Helvetica;
+	font-size: 12px;
+	font-weight: 500;
+	letter-spacing: 0;
+	line-height: normal;
+}
+
+.box .date,
+.mode {
+	height: 25.94px;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	min-width: 71.17px;
+	position: relative;
+	border: 1px solid var(--blue);
+	border-radius: 0.5rem;
+}
+
+.button-wrapper {
+	border: 0px none;
+	height: 36.14px;
+	overflow-x: hidden;
+	width: 132.28px;
+	display: flex;
+	justify-content: end;
+}
+
+.button {
+	border-radius: var(--borderRadius);
+	padding: 0.3rem 1rem;
+	background-color: var(--blue);
+	color: var(--White);
+	border: 1px solid var(--blue);
+}
+
+.heading {
+	color: var(--blue);
+	display: flex;
+	align-items: center;
+	font-family: "Plus Jakarta Sans-Medium", Helvetica;
+	font-size: 2rem;
+	font-weight: 700;
+	letter-spacing: 0;
+	padding: 1rem;
+	line-height: normal;
+	width: 100%;
+	white-space: nowrap;
+}
+
+.all {
+	margin-top: 3rem;
+}
