+++ src/modules/Dashboard/modules/Hackathon/services/HackathonApis.ts
@@ -1,10 +1,45 @@
 import { AxiosError } from "axios";
-import { privateGateway, publicGateway } from "@/MuLearnServices/apiGateways";
-import { dashboardRoutes, onboardingRoutes } from "@/MuLearnServices/urls";
-import { HackList } from "./HackathonInterface";
+import { privateGateway } from "@/MuLearnServices/apiGateways";
+import { dashboardRoutes } from "@/MuLearnServices/urls";
+import { HackList } from "./HackathonInterfaces";
 import { SetStateAction } from "react";
-import { ToastId, UseToastOptions, useToast } from "@chakra-ui/react";
+import { ToastId, UseToastOptions } from "@chakra-ui/react";
 import { Option } from "@/MuLearnComponents/FormikComponents/FormikComponents";
+
+export const getHackathons = async (
+    setData: React.Dispatch<SetStateAction<HackList[]>>
+) => {
+    try {
+        const response = await privateGateway.get(
+            dashboardRoutes.getHackathons
+        );
+        const defaultForm: any = response?.data;
+        setData(defaultForm.response);
+    } catch (err: unknown) {
+        const error = err as AxiosError;
+        if (error?.response) {
+            console.log(error.response);
+        }
+    }
+};
+
+export const getOwnHackathons = async (
+    setOwnData: React.Dispatch<SetStateAction<HackList[]>>
+) => {
+    try {
+        const response = await privateGateway.get(
+            dashboardRoutes.getOwnHackathons
+        );
+        const defaultForm: any = response?.data;
+        setOwnData(defaultForm.response);
+        console.log(defaultForm.response);
+    } catch (err: unknown) {
+        const error = err as AxiosError;
+        if (error?.response) {
+            console.log(error.response);
+        }
+    }
+};
 
 export const getFormFields = async (
     setFormData: React.Dispatch<React.SetStateAction<string>>
